sysctl: Limit sched_upmigrate/sched_downmigrate tunables range
The sched_upmigrate/sched_downmigrate tunable values are
accepted in percentage. Hence don't accept any values greater
than 100. The sched_upmigrate is also used to check if a CPU
is overutilized or not. Accepting a large value for this tunable
makes a heavily loaded CPU not getting classified as overutilized.
Since the CPUs are not overutilized, the higher capacity CPUs
does not offer help to the heavily loaded lower capacity CPUs.
The no-hz idle balance kicks are also not sent to the CPUs with in
the same cluster. This results in starvation.

Change-Id: I83c3caeeb61f05de5b74b41c22702cb895072fad
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: engstk <eng.stk@sapo.pt>